testes-junit/src/gleidev/com/junit/servicos/LocacaoService.java
package gleidev.com.junit.servicos;

import gleidev.com.junit.entidades.Filme;
import gleidev.com.junit.entidades.Locacao;
import gleidev.com.junit.entidades.Usuario;

import java.util.Date;

import static gleidev.com.junit.utils.DataUtils.adicionarDias;

public class LocacaoService {

    public Locacao alugarFilme(Usuario usuario, Filme filme) {
        Locacao locacao = new Locacao();
        locacao.setFilme(filme);
        locacao.setUsuario(usuario);
        locacao.setDataLocacao(new Date());
        locacao.setValor(filme.getPrecoLocacao());

        //Entrega no dia seguinte
        Date dataEntrega = new Date();
        dataEntrega = adicionarDias(dataEntrega, 1);
        locacao.setDataRetorno(dataEntrega);

        //Salvando a locacao...
        //TODO adicionar método para salvar

        return locacao;
    }
}